Pass SavedFilter initial parameters as JSON & relocate test
